move models definitions
