+++ backend/case_study/admin.py
@@ -1,3 +1,10 @@
 from django.contrib import admin
+from . import models
+from event.admin import EpistolaryEventInline
 
-# Register your models here.
+
+@admin.register(models.CaseStudy)
+class CaseStudyAdmin(admin.ModelAdmin):
+    fields = ["name"]
+    inlines = [EpistolaryEventInline]
+    extra = 0
