+++ backend/core/models.py
@@ -1,3 +1,22 @@
 from django.db import models
 
-# Create your models here.
+
+class Field(models.Model):
+    certainty = models.IntegerField(
+        choices=[
+            (0, "uncertain"),
+            (1, "somewhat certain"),
+            (2, "certain"),
+        ],
+        default=2,
+        help_text="How certain are you of this value?",
+    )
+
+    note = models.TextField(
+        null=False,
+        blank=True,
+        help_text="Additional notes",
+    )
+
+    class Meta:
+        abstract = True
