+++ backend/letter/admin.py
@@ -1,3 +1,14 @@
 from django.contrib import admin
+from . import models
 
-# Register your models here.
+
+class LetterMaterialAdmin(admin.StackedInline):
+    model = models.LetterMaterial
+    fields = ["surface", "certainty", "note"]
+
+
+@admin.register(models.Letter)
+class LetterAdmin(admin.ModelAdmin):
+    inlines = [
+        LetterMaterialAdmin,
+    ]
