+++ backend/person/admin.py
@@ -1,3 +1,12 @@
 from django.contrib import admin
+from . import models
 
-# Register your models here.
+
+class PersonNameAdmin(admin.StackedInline):
+    model = models.PersonName
+    fields = ["value", "certainty", "note"]
+
+
+@admin.register(models.Person)
+class PersonAdmin(admin.ModelAdmin):
+    inlines = [PersonNameAdmin]
